Refactoring to include src folder and updated tests
